+++ site/dat/docs/Changelog.md
@@ -13,6 +13,7 @@
  - add `content-encoding` request- and scenario-level option to JMeter
  - change routing part for BlazeMeter reports opened in browser
  - fix log verbosity for HTTPError
+ - send `additional-classpath` jars into cloud when preparing a cloud test
 
 ## 1.7.2 <sup>13 oct 2016</sup>
  - fix keep-alive processing in Gatling